feat: 메인페이지 hook 분리 #92
